requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gunicorn
requests
psycopg2-binary
python-dotenv 
whitenoise